requirements.txt
coloredlogs==15.0.1
deep-translator==1.11.4
python-dotenv==1.0.0
selenium==4.18.1
tweepy==4.14.0
